Correctly interpolate old solution to new elements
The code now refines elements in the reference space, so that the
coordinates of the nodes in reference space can be used to evaluate the
basis functions. This way, the basis functions can be combined with the
old function values to get an interpolant for the new element. The
isentropic vortex case now works successfully while splitting a few
elements.