Vectorized analytic flux for Euler2d
The interior flux function for Euler2d has been vectorized and
arithmatically simplified to minimize repeated calculations. Many things
in the code have been commented out, including calls to calculating face
and boundary face residuals, so as to make the code run to completion.
These will be slowly uncommented as they are vectorized.