Add nonlinear solve to Trapezoid method
Previously, Trapezoid used linearization to solve the nonlinear system
that results from the implicit method. However, on the pendulum case,
this was observed to be only first order accurate. To improve this, the
linearization was replaced with Scipy rootfinding. The integrator is now
second order accurate for the pendulum case.