Fix Pytest breaking on failure, add progress bar
Resolved an issue with Pytest failing after an end-to-end test failed.
This was because it exists the test after the assert fails, so never
runs the command to move back to the original directory. This command
was moved to be after the 'yield' statement in the fixture for the test,
since everything after 'yield' is the teardown for the fixture.

Also, added progress bar outputs. This output option (off by default)
replaces the per-iteration information output with a colored progress
bar that gives the percent of the simulation that has completed in
physical time (computed using the current time over the final time).
All of the end-to-end tests now have this enabled to decrease the amount
of text spam from a failed test case.